Introduce clock_gettime_ns()
Summary:
The clock_gettime() API is slow, as it uses references. Let's provide a
clock_gettime_ns() API that avoids a lot of unnecessary math and simply returns
an integer. Also, avoid the overhead of CLOCK_THREAD_CPUTIME_ID handling by
optimizing the call to the underlying implementation if the requested clock is
not CLOCK_THREAD_CPUTIME_ID.

Reviewed By: paulbiss

Differential Revision: D5734060

fbshipit-source-id: 0d5cd1ef55b9c7d12d442e165a2561bcc4c336e2