Store PackedArray size class in 8-bits, make room for [d/v]array bits
Summary:
Use 8-bits in the HeapObject header (aux16) to store an enum indicating whether
an array is a darray, varray, or neither. However, PackedArray already uses
aux16 to store its size class. The size class can never exceed 8-bits, so
instead store the size class in the upper 8-bits of aux16, and use the lower to
store the [d/v]array bits. Adding unions with 8-bit members to HeapObject causes
a sizable regression (probably because of aliasing rules), so do the bit
shifting ourself (this doesn't generate any worse code).

For now, just store "not d/varray" in the bits. Later diffs will start to set it
properly.

Reviewed By: mxw, alexeyt

Differential Revision: D5746758

fbshipit-source-id: bc11a06ea36d5143630075bcd2f73a0dc3491561