Enable HHVM intrinsics extension consistently in repo-auth mode
Summary:
The EnableIntrinsicsExtension option wasn't enabled for HPHPc or HHBBC,
which meant that HHBBC wouldn't consider calls to any builtins within the
extension as actual calls to builtins. This is mainly harmless, but its better
to have HHBBC have the same view of extensions as during runtime. Fix this by
loading and storing EnableIntrinsicsExtension in the repo global-data. Also
enable it for HPHPc invocations in the test runner.

Reviewed By: markw65

Differential Revision: D5749357

fbshipit-source-id: 6d16c757eb67210631b10200ff44ee9e819e1ce4