AArch64 Fix gold linker to work on AArch64
Add variables to control usage of the gold linker options --icf=
and ensure the final hhvm binary is not marked in the ELF header
with an executable stack section.

Assembly files, even empty, are now marked with the directive that
prevents them from having an executable stack. Besides locking down
security somewhat, it also brings total runtime of the unit test
suites back in line with the old ld.bfd linker (which does not
mark the ELF as having an executable stack section).

The --icf=all option currently works on AArch64 Ubuntu but not
AArch64 CentOS. When cmaking the hhvm project on CentOS, one can
-DICF="-Wl,--icf=none" to prevent a broken binary. CentOS currently
crashes with --icf=all in boost.*program_options.* when parsing the
the command line parameters. Who knows what else is broken due to
ICF after that. There are 77 symbols ICF'd from/into boost's
program_options, so using the --keep-unique SYMBOL gold linker
option is fairly unreasonable. I suspect but haven't proven that
the directions for building and installing boost on CentOS can be
modified to prevent ICF from folding boost into HHVM objects.