generate Tick even when server not connected
Summary:
To deal with a lost server, we will poll the lock_file once a second to see if the monitor has been restored externally.

Prior to this diff, the Tick message was only generated when the server was connected. This diff makes a Tick also be generated even when the server isn't connected.

Reviewed By: arxanas

Differential Revision: D5595602

fbshipit-source-id: caa2c9d03b0cfc2c0d2a4fbf660173170825151b