Delete retry_if_init
Summary:
The server might have been in an "initializing" state when someone tried to connect to it. So the client commands like `hh check` had a command-line option `retry-if-init`, and this was passed as a boolean into ClientConnect.

Now we have a monitor and purgatory clients and this retry-if-init is no longer needed. So I'm deleting it.

I'm retaining a command-line option but just ignoring it. Pretty much everyone out there still passes a `retry-if-init` flag, and I don't want them to break.

Reviewed By: alexchow

Differential Revision: D5570446

fbshipit-source-id: 8ca5d6cc4cba55a4655617de5d60d2584ad34519