refactor do_initialize
Summary:
We need to connect in two contexts...

(1) in response to a do_initialize request, in which case we give an initialize-response back to the client in case of success and report things like initialization failure via exceptions for the client to restart;
(2) in response to losing the connection and then needing to regain it, in which case we don't give an initialization response back to the client, and failure has to be dealt with by us not the client.

This diff refactors out "do_initialize" (which handles part 1) from "connect" (which is the common connection routine called by 1 and 2).

Reviewed By: arxanas

Differential Revision: D5616182

fbshipit-source-id: f8d1574447dccedca7151e53ffadc5827e9c3db6