use ClientConnect{autostart=true}
Summary:
Prior to LSP, the way it worked was that Nuclide did "hh start" to start the server (which calls ClientStart), and then it launched "hh ide" (which calls ClientConnect{autostart=false}).

Why did it do that instead of just ClientConnect{autostart=true}? Because the startup process (be it via "hh start" or via "autostart=true") wasn't suitable for use in a service - it send output to stdout and stderr, and terminated abruptly in some cases. By having Nuclide call "hh start" first it was easy to swallow all that irregular output. Afterwards, "hh ide" would only have clean behavior.

In moving to LSP we just simply copied this pattern: the LSP server first does ClientStart and then it does ClientConnect{autorestart=false}. It took a fair bit of refactoring on the ClientStart to get rid of most of the cases of irregular output and abrupt exits.

But now that we've cleaned it up, well, there's no point anymore. We might as well just call ClientConnect{autorestart=true}. That's what this diff does.

(There is a slight difference in behavior: ClientStart is slightly more aggressive about force-killing the old server if the monitor isn't immediately responsive, which it interprets as "stuck" or "crashed", although it should also recognize the possibility that the monitor is simply in the process of starting up. The ClientStart behavior isn't great).

Reviewed By: alexchow

Differential Revision: D5620609

fbshipit-source-id: 7a7a7e6ad2703af812d387bc2f92188b8b2dd01e