Add support for namespaces
Summary:
Adds namespace environment to the visitor environment.

We handle namespaces really weirdly in Hack — we have a namespace elaborator that runs right after parsing, but only elaborates the toplevel namespace entities. The rest of namespace elaboration occurs at naming.

In order to avoid having to run all of naming, I just update the namespace elaboration manually for each type of dependency we add. It's pretty straightforward to do this since we only need to elaborate the namespaces when we add a dependency and we know exactly the type of dependency we wish to add.

I would have preferred if I could write a "visitor_with_namespace_env" generic class to do this for me, but this seems really difficult(we'd need some way to treat classes in OCaml as first class entities, almost like a functor between classes. I tried converting the classes to modules, but then we run into other woes). This was the simplest and cleanest solution for now.

Reviewed By: dabek

Differential Revision: D5654909

fbshipit-source-id: cfb90a2cd7fcea6d296c8b820ef90514873e0829