Move Some HHBBC::Options into RuntimeOption
Summary:
There are certain `HHBBC::Options`, i.e. `HardTypeHints`/`HardReturnTypeHints`, that are currently only used in repo mode, and is stored in `Repo::GlobalData`. Now that we are enabling HHBBC for non-repo mode, we need to expose these options to the user and make them accessible without querying `Repo::GlobalData`. There might be more options that need to be moved into `RuntimeOption`, but I lack the knowledge.

In repo mode, we invoke `hphpc`, `hhbbc`, and `hhvm` in listed order. Some options, especially `HHBBC::Options` we set to `hphpc` are serialized into the repo, and later reads out in `hhbbc` and `hhvm`. These options should override or force set/unset any runtime options specified in the `hhvm` step so that `hhbbc` and the runtime are on the same page.

In non-repo mode, we invoke `hhvm` directly, and we will either use `RuntimeOptions`, or, in rare cases, we uses `HHBBC::Options` default values defined in `hhbbc/options.h`.

This diff makes the following changes

* Added `RuntimeOption::EvalHardTypeHints`,removed `HHBBC::Options::HardTypeHints`. Default was true for repo mode, and false for non-repo mode. Default is remained.
* `HHBBC::Options::HardReturnTypeHints` is removed, and we test `RuntimeOption::CheckReturnTypeHints >= 3` instead. Default was `true` for repo mode (the HHBBC option) and 2 for non-repo mode (the runtime option, means `false`). The default value is remained. Note that, in repo mode, `hphpc` and `hhbbc` only care about whether the option is smaller than 3 or not. Therefore, in the `GlobalData`, we only record a boolean value, and we save to/read from `GlobalData`. But in the last step in repo mode, when we invoke `hhvm`, we don't read the value from global value, but read it from the command line since this time a boolean flag is not enough, and we need a precise value for this option.
* `HHBBC::Options::CheckThisTypeHints` is removed, and we test `RuntimeOption::CheckThisTypeHints` instead. The default was, and is, false.
* `HHBBC::Options::DisallowDynamicVarEnvFuncs` is removed, and we test `RuntimeOption::DisallowDynamicVarEnvFuncs == ON` instead. The default was based on `EnableHipHopSyntax` for the runtime option. The HHBBC version of this option is removed, since HHBBC should always follow HPHPC on this option.
* `RuntimeOption::SoftClosureReturnTypeHints` is removed since it is not used anymore.
* `HPHP::Option::UseHHBBC` is removed, and we test `RuntimeOption::EvalUseHHBBC` instead. The default was/is `!getenv(HHVM_DISABLE_HHBBC)`.
* Change test case options accordingly.

Reviewed By: markw65

Differential Revision: D5682232

fbshipit-source-id: 5882dc3cfb0f20d313b0170de26786649f9b5afe