Fixing regexp for trivia lowering
Summary:
The lowerer (still) uses regular expressions to find interesting information in `HH_FIXME` and `HH_IGNORE_ERROR` pragmas (i.e. the error code that's covered). These regular expressions are less permissive than the trivia parser. This broke:
```
/* HH_FIXME[1234: Some remark */
```
due to the missing `]` and
```
/* HH_IGNORE_ERROR [1234]: Some remark */
```
due to the space before the `[`. This diff is a quick fix. This information should really be extracted structurally from the trivia parser.

Reviewed By: hubyrod

Differential Revision: D5764109

fbshipit-source-id: 793018492f25b6075bbac4ffbba38a84559c2e59