Add the original node being lowered to the coroutine context
Summary:
In the previous diff I mentioned that there is a bug in the lambda analyzer whereby it does not correctly compute outer variables of an anonymous function. To fix this bug, the lambda analyzer must know whether it is analyzing a lambda or an anonymous function. However, right now the entrypoint to the analyzer takes the parents, the body and the parameter list; it does not take the original node of the function being analyzed! It would be smart to pass in that node (and the parents) and extract the body and parameter list from it.

This means that the original node needs to be passed around the coroutine lowering code; the easiest way to do that is to simply add the original node to the coroutine context object that is already being passed around.

In the next diff I'll fix up the lambda analyzer to take the original node being analyzed, and then I'll fix the bug.

Differential Revision: D5760137

fbshipit-source-id: 28cbeb872d139a85498e3cf87429956fd631c3c1