Fix entrypoint to lambda analyzer so that it takes node being analyzed
Summary:
The entrypoint to the lambda analyzer had a dumb entrypoint; it took the parent list, the parameters, and the body, but not the actual node being analyzed, which is (1) weird, and (2) gets analysis of anonymous functions wrong.

The anonymous function bug will be fixed in the next diff; this just gets the entrypoint correct.

 

Differential Revision: D5760238

fbshipit-source-id: 9500deb2abb8cf6faff75ba628e25ac6ead2cef0