Convert system to use Typing_deps dependencies
Summary:
This changes the system to use the same system as Typing_deps. This doesn't guarantee that we'll use Typing_deps to actually implement storing and USING the dependencies, but right now I need it to be consistent so I can compare test files to make sure we always return a superset of what our original dependency system returns.

This diff makes it so that hh_single_type_check --dump-toplevel-deps and --dump-bazooka-deps both return consistently formatted results.

Reviewed By: dabek

Differential Revision: D5729754

fbshipit-source-id: 344a1d2d6c654d8c95a916688ad06fdf9e3e39cb