Keeping path in source text structure
Summary:
There are a few places that invoke `failwith` to bail out. When this happens in a run of `hh_server` on a large repository, it would be nice to know where the error happened. Instead of
```
Exception: Failure("(10006-10006) Expected token < but got (")
```
it would be nice to see the offending file, i.e.
```
Exception: Failure("(10006-10006) Expected token '<' but got '('\n  in /path/to/your/file.php\n")
```

To do this, the path to the source file needs to be available. `Full_fidelity_source_text` didn't carry the path with it. Note that the cost of this is a pointer per open file (so invariant to our high velocity `lexer` creation).

Reviewed By: ericlippert, alfa07

Differential Revision: D5764102

fbshipit-source-id: d73b5c4c3b580fa01437ee19faa0ca91e3755684