Make `parses_without_error` exception-safe
Summary:
The Full Fidelity Parser uses a combinator `parses_without_error`, to do some speculative lookahead parsing. Given
```
<?hh // strict

class C {
  public function foo(): void {
    bar((tuple(4,2)));
  }
}
```
made the FFP break with a `Failure`. This is because too much of the argument of `bar` looks like a lambda, which delegates to the declaration parser and throws there.

Admittedly, this feels like a bit of a hack, but it fits the reasonably expected semantics of `parses_without_error` to catch such `Failure`s and translate them to `false`.

Reviewed By: hubyrod

Differential Revision: D5764293

fbshipit-source-id: 54e2ddc50e7bb8430c3f39aa2f81ce13d6cfdccd