Don't throw types of XHP attribute declarations on the floor
Summary:
Fix a bug in the full fidelity parser where, if the input has keywords as attribute names, the parse tree would not contain the attribute's type.

The key removal in this diff was added in D5222552, but considering that didn't state what it was for and it is known-broken as is, this diff removes it.

Reviewed By: ericlippert

Differential Revision: D5766631

fbshipit-source-id: eaca412d7208372f73aaf3fbefba28856a5d6cca