Get rid of the array_do macro
Summary:
array_do was used to expand to one of three different calls, which
were implemented by 2 different template functions.

The two shared calls (set and elem) don't really want the same
implementation, so get rid of the macro and the templates, and just
write out the three separate functions.

This allows us to avoid some copy-on-write in array_do_set and
array_do_newelem by std::moving the base into the helper functions,
and updating the base before returning in the successful cases.

I also removed some redundant "weird key" checking; array_elem already
takes care of unioning in InitNull when the type of the element being
accessed can't be determined exactly (which is always the case with
weird keys).

Reviewed By: ricklavoie

Differential Revision: D5771492

fbshipit-source-id: b9b94f2611a49fd4f384894d626da8654e796818