Simplify GMP::unserialize.
Summary: The GMP class doesn't registerNativePropHandler, have a magic __set method, or declare any properties (so all its properties are dynamic). That means that using o_setArray in GMP::unserialize is overkill; we can just take the Array of unserialized properties and setDynPropArray it onto this_. Add a unit test to verify this works as intended.

Reviewed By: markw65

Differential Revision: D5774731

fbshipit-source-id: 8fccbac0085536c2a4375238e8f215be1c2ddf92