Constant fold foldable builtins
Summary:
Fold FPush...FCall calls to foldable builtins. We can propagate the
result at optimization time, but unlike FCallBuiltin, we need to use dce
to actually eliminate the calls.

This should work for any side-effect-free call with a known result -
but we don't currently detect side-effect-free calls; I'll follow up
with a diff to do so.

Reviewed By: ricklavoie

Differential Revision: D5763153

fbshipit-source-id: d21e7ff00fcb360cda01b2de28d0a0a3e85a78ea