Scan Sweepable, kill EphemeralPtrWrapper<T>
Summary:
The linked list of Sweepable objects and resources consists of a head
node in MemoryManager, and Sweepable nodes embedded in the live objects
and resources in the heap. Currently we scan this entire list as part of
the root set, however it should only be necessary to scan the head:
Each pointer to a Sweepable is actually an interior pointer to either an
ObjectData or ResourceHdr allocation, and the GC supports interior pointers.

This eliminates the last use of EphemeralPtrWrapper, the last use of
type_scan::Scanner::enqueue(), and the last use of Scanner::m_ptrs, which
is a list of pointer *values* for which we don't know the address and thus
cannot update later, when an object is copied. m_ptrs is not to be confused
with m_addrs, which contains *addresses* of pointers, thus allowing updates.

EphemeralPtrWrapper, and pointers with unknown owners, were generally
unfriendly to tools built with HeapGraph, because it makes it difficult to
identify why certain objects are root-reachable. Scanner & HeapGraph cleanup
will be in a followup commit.

Differential Revision: D5763010

fbshipit-source-id: 01e57e37eb18700c228bc70bcf5ef50d2f3fcd39