Drop trailing commas at the start of a formatting range
Summary: The subchunk variant `Comma` represents a trailing comma that was added because its associated split was broken on. Since we don't have source mapping information for `Comma`s (since there may have been no trailing comma in the original source), we only want to print them when the atom preceding the trailing comma is printed. If a `Comma` is at the beginning of the range, then the atom preceding it was not included, and neither should the `Comma`.

Reviewed By: arxanas

Differential Revision: D5772427

fbshipit-source-id: 0757e64b4133b135a6bca93a0f659eaabf4e2305