Unbreak APCObject fast init in debug mode.
Summary:
We had no unit test coverage for APCObject fast init, because in order to trigger the optimization, you have to have Server.APC.AllowObject=true, apc_store and then apc_fetch an object of a class with no constructor, and the class has to satisfy classHasPersistentRDS().
Our APC unit tests that set Server.APC.AllowObject=true were declaring classes in the same file they were instantiating them, so classHasPersistentRDS was always false and the optimization never triggered.
Add a test that triggers the optimization, and fix the bogus assert that snuck in due to the lack of test coverage.

Reviewed By: markw65

Differential Revision: D5784248

fbshipit-source-id: faa24d819b616e8c07f3b38a767932c72a2c6f7f