Fix CompactVector for types that can't be memcopied
Summary:
I tried to be clever, and avoid the overhead of std::move/destroy,
thinking we'd never hit a real world case where it matters. But I just
stuck a FuncAnalysis in a CompactVector and discovered that
std::unordered_map *does* require you to std::move it (after hours of
debugging).

Pretty sure swtaarrs warned me...

Reviewed By: swtaarrs

Differential Revision: D5780441

fbshipit-source-id: 9f2aaf95276539ac22390593a5522f086e830964