JIT support for one-bit refcounting
Summary:
Most of the changes in the jit were just finding all the uses of
FAST_REFCOUNT_OFFSET and moving things into helpers that do different things
depending on the value of one_bit_refcount. Every operation that directly
touches a reference count is now in code-gen-helpers.*.

Along the way, I decided to PrivateValue/SharedValue to
OneReference/MultiReference, and kill InitialValue in favor of OneReference
everywhere. This gets rid of a few conditionals in countable.h and simplifies
some of the code I modified in the jit. I also audited all uses of
(Maybe)?Countable::isRefCounted() and found a few incRefCount() calls that
could be rawIncRefCount().

Reviewed By: edwinsmith

Differential Revision: D5752621

fbshipit-source-id: 8d991d10ea23e166112ffaeec5a731cbc22654fd