Propagate lambdas' ORIGINAL method/function/anonymous function/lambda positions through to the lowered EditablePositionedSyntax
Summary:
= Background

We perform the following lowering:

{ `MethodishDeclaration`, `FunctionDeclaration`, `AnonymouFunction`, `LambdaExpression` } --Lowered into-> State machine

= This change

This diff carries the positional information from the coroutine definition into the generated state machine. This means that, with this diff, the generated state machine **as a block** will be positioned. So, I'm **not** guaranteeing that every statement in it is positioned, but I am guaranteeing that the entire function's span is positioned (e.g., "This function starts on line 5 and ends on line 12").

= Why?

So, we can //currently// step **onto** individual statements in the coroutine, but we can't step **over** them. This is because HHVM apparently uses the function's positional information to determine what it means to step "over" something. This diff fixes that. Consequently, with this change in place, we have pretty good coroutine debuggability. You can "kind of" step over statements within a coroutine.

"Kind of": We're emitting a ton of empty .srclocs. The debugger gets confused on these, and stops on them. This is useless, since they correspond to generated code. In the next diff, I fix this.

**Note:** This diff adds some complexity, as we have to carry through the original method/function/anonymous function/lambda **as a syntax node** (so, `Full_fidelity_editable_positioned_syntax.t`). This node contains the positional information. This is different from everything else we've carried through in the framework, which we've carried through as the actual type (e.g., `Full_fidelity_syntax_type.methodish_declaration`). I don't like this, but I'm hoping to punt it to a later diff, because I think there's a lot of refactoring that needs to be done in this area, and it would just add (a lot of) noise to this diff if I tried to do it here. I'm happy to look more into doing it here if the reviewers would prefer that, though.

**Overall design:** https://fburl.com/coroutine-debugging

Reviewed By: ericlippert

Differential Revision: D5743169

fbshipit-source-id: e4b6006d052a49232acc0418db638d92ef39a308