Do not emit .srcloc instructions whose positions are Pos.none
Summary:
This is a tiny change in emit_pos.ml. With this change, any `.srcloc`s that have a position of `Pos.none` will **not** be generated in the lowered HHAS.

This is important for the debugger: The debugger will stop on Pos.none positions, and get quite confused. In coroutines, these lines correspond to generated code, that we're not interested in exposing to the user. By omitting these locations, we can hide them from the debugger.

**Overall design:** https://fburl.com/coroutine-debugging

Reviewed By: andrewjkennedy

Differential Revision: D5723993

fbshipit-source-id: 8380e54d153a30a78e4c606058db7bed02210a97