Remove split between leading binary operator and right-hand side
Summary:
There should never be a line break between a leading binary operator and its right-hand side, and the way to express "never break" in hackfmt's IR is to put content into a single Chunk. Previously, however, range-formatting only worked at Chunk boundaries, which meant that we'd duplicate content when a formatting range ended immediately after a pipe operator (which happens when the pipe operator was formatted as a trailing operator and hackfmt attempts to switch it to leading. See test case binary_expression_range_formatting.php). To deal with this, we added a split between the operator and the right-hand side, even though we didn't actually want to permit breaking there.

Since then, we have added support for range formatting at non-Chunk-boundaries in order to support as-you-type formatting, so the conditions that required us to add this Split are gone. We can now remove it.

Reviewed By: arxanas

Differential Revision: D5791090

fbshipit-source-id: 6526d5df8e040fe46b6e37c6672be6f8d440bea0