lost server part 1 - basics
Summary:
This stack of changes is a rethink of how to handle the connection between the hh_client_lsp server and hh_server. This stack is primarily related to how and when we deal with a lost connection to hh_server.

OVERVIEW

The lifetime of hh_client_lsp will be completely decoupled from that of hh_server. The hh_client_lsp server will run for the duration that Nuclide/VSCode wants it. It will display the current status of hh_server (stopped / initializing / ready), and will react appropriately to changes in the current status at any time. At times when hh_server is stopped, it will offer the user a “restart” button. Note that LSP has the initial handshake deal with initialization-failure. We will not leverage this part of the protocol, since we deal with initialization-failure at any stage (not just at the beginning). Instead hh_client_lsp will always report immediately that LSP initialization succeeded.

SOME GUIDING PRINCIPLES:

* if the editor has dirty state (i.e. unsaved edits) and we've lost hh_server, then hh_server has NO WAY of getting that state back itself, and the client will NEVER send that state to a running+working LSP server. The only way to get it back is if clientLsp.ml actually exits, and we trust the client to restart it.
* the client has no logic in its triggers of when to restart; it simply restarts the LSP server automatically. If we want fancy logic (e.g. reconnect when hh_server is started externally, or reconnect when focus shifts and an LSP verb happens) then we have to do it ourself inside a still-running LSP server.

HOW MIGHT WE LOSE A SERVER CONNECTION?

* if someone takes over the persistent connection
* if someone does 'pkill hh_server' or if server crashes
* if someone does 'hh stop' or 'hh restart' (or even 'hh start' in some race condition situations)
* if .hhconfig changes and hh stops, but will restart once it has determined that a currently-underway source control operation has finished.

WHAT WILL WE DO?

* put up a dialog saying that hh_server has stopped, with a [Restart] button
* if they dismiss the dialog, we'll put up an actionRequired indicator telling them that hh_server needs to restart

WHAT TRIGGERS A RECONNECT?

* if user clicks the [Restart] button on the dialog
* in case of stolen persistent connection, we'll reconnect automatically upon any LSP method
* in non-stolen cases, we'll reconnect automatically if the hh_server lockfile reappears.

Reviewed By: alexchow

Differential Revision: D5620613

fbshipit-source-id: 7faa270ae64078bd672c8b46a9424e10935e1741