lost server part 2 - triggers
Summary:
The codebase already supported one trigger for regaining a lost server -- the trigger that, if any LSP request/notification should arrive, then regain the server. (This was used in cases where the server had been lost due to a different window stealing the persistent connection).

This diff adds support for two new triggers:

(1) When we enter the Lost_server state, we now put up a dialog request. If the user clicks on the button in that request, well, this is the trigger. (If the user dismiss the dialog request then we instead put up an actionRequired indicator, if it's supported by the client).

(2) The other trigger is if ever the monitor lock_file comes into existence. If this trigger is enabled then we poll for the lock_file once a second while in Lost_server state.

Reviewed By: alexchow

Differential Revision: D5620618

fbshipit-source-id: f4592bb7fbf243846fca6703405e180ccd3d01c8