lost server part 3 - dirty files
Summary:
Imagine if there are unsaved editor buffers. LSP client will never re-send the contents of these buffers to an existing LSP server. So: if hh_server crashes (losing its mirror copy of the editor buffers), the only way forwards is for “hh_client lsp” itself to terminate, and rely upon Nuclide/VSCode to relaunch it and thus re-send the contents of the editor buffers.

This code keeps track of whether there are dirty buffers, so it can later make the decision of whether to terminate LSP or just reconnect.

Reviewed By: alexchow

Differential Revision: D5620621

fbshipit-source-id: bcfb47858cc43e12d5b4ee92352bed65019042d6