lost server part 6 - never autostart
Summary:
Previously, Nuclide would auto-start hh_server if it wasn't already running. This diff changes it so that Nuclide will never spontaneously start hh_server; it will happily connect to a running hh_server whenever there is one, and it will offer the user a convenient button to start the server should it be down, but it'll never actually start the server automatically.

Why?

Well, our expectation is that hh_server is always running. That's what a proper machine is like, and it's good because it means it's instantly available for everything. There are a few circumstances where it's not running:

(1) if the user did "hh stop" or "pkill". In this case I think the user does *NOT* expect that their Nuclide will spin the server up under their feet. They probably expect to start the server manually.

(2) if there was some kind of really serious crash. In this case that underlying problem should be addressed for sake of future efficiency, and I think it's wrong to paper over it.

There's one other impact of this change. If our persistent connection goes down, then we normally use the presence of the lockfile to judge when we should reconnect to the server. But imagine if the user did "hh stop" and the persistent connection went down but then we examined the lockfile before it had been removed. Then we'd think it was time to reconnect. If we had autostart then we'd end up starting up the server even though the user had explicitly shut it down. With this change, we won't.

Reviewed By: alexchow

Differential Revision: D5620649

fbshipit-source-id: 774318309136a9dc430fd2f41db195b5711b1d37