don't report "initializing..." if it connects instantly
Summary:
If we end up connecting to hh_server instantly, then we shouldn't pop up the "initializing..." notification.

This diff implements that. It no longer displays the notification at the moment we made the attempt to connect to hh_server. It only displays it if a second goes by (i.e. we get a "Tick" message) and we haven't yet heard a "hello" back from the server.

Reviewed By: arxanas

Differential Revision: D5620657

fbshipit-source-id: 3dfdd9048da29968a2d4193e6fa687801d656fe0