don't flicker 'Restart' if it'll be brief
Summary:
Imagine you tweak .hhconfig to cause an server restart, or you do 'hh restart'.

Previously it would flicker up "hh_server is stopped [restart]" for a moment as it entered the Lost_server state, and then on the next tick it would see that the lockfile is present and so reconnect.

This diff eliminates the flicker. If it loses the server but sees the lockfile is present, then it will avoid putting up the dialog, and will instead go straight to reconnection.

Reviewed By: alexchow

Differential Revision: D5640025

fbshipit-source-id: 5b2bde481f29bf01a937c6dfb728106c2767c7c8