Make better use of local-equivalency
Summary:
We didn't allow equivalency when the local might be uninit - but many
uses of it don't care, and the ones that do can check.

In addition, when we refine local types based on assertions -

  if ($x instanceof X) { $x must be an X here }

we can refine all the equivalent locals, rather than just the one
being asserted.

Finally doing this will make it possible to simplify or eliminate a
lot of the group helpers in interp.cpp (eg we can get rid of "cgetl;
jmp", because the jmp opcodes can see that the top-of-stack is
equivalent to a set of locals anyway). Will do a follow up.

Reviewed By: ricklavoie

Differential Revision: D5795643

fbshipit-source-id: a22d56a9cee976af9bb85c432833e0c261a94395