Simplify/generalize some groups
Summary:
Now that stack locations have a reliable equivLoc, we can generalize
or eliminate the groups that started with CGetL. By doing so we'll
match more patterns (eg CGetQuietL;... or CGetL; Dup; ... or CGetL;
CGetL2; ...).

I'll do a follow up to use a flag to indicate that a stack element is
a dup of the one below, which will let us get rid of the groups
beginning with Dup.

Reviewed By: ricklavoie

Differential Revision: D5798343

fbshipit-source-id: cbaad2122519f70bb4e359e32e2cbd54b2f6b1cc