Preserve original body of the lambda
Summary:
This is step 1 in supporting lambdas in default values for parameters.

Previously closure conversion process replaced original body of the lambda with some special values that indicated if anonymous function was using 'use' langugage construct. As a consequence during emit we cannot support lambdas in default parameter values because initial body was lost. This diff removed usage of function body as indicator in favor of explicit table.

Reviewed By: hubyrod

Differential Revision: D5788738

fbshipit-source-id: 69c476d7458881e7c8c16a0ad682aaa1733db64b