Informant uses source control aware mode
Summary:
Mercurial has an Scm_aware mode that sends a notification to
the subscribers when the merge base has changed. The new merge
base is still specified as a hg revision (which we still must map to a
corresponding SVN revision to compute its distance).

This diff adds support for that mode. When using this mode, our Watchman
client also adds the "Drop_changes" option, which means we won't get the set
of files changed when moving between merge bases (i.e. if you move from trunk revision
A to B, watchman will only send us the State enter, State leave, and Changed_merge_base
notifications, and not the files that changed between A and B). Since we won't be using
Scm_aware mode in the typechecker itself, I can't think of a reason why we would want
those file changes anyhow.

This diff also uses this new mode in the Informant. Previously, when we see State_enter
and State_leave events, we would kick off asynchronous work (such as prefetching the
saved states and shelling out to hg to map the hg_rev to an svn_rev). And then when the
mapping to svn_rev is eventually ready, we would "form_decision" on the notification
to decide to take any action (the only action we actually take is Restart_server on a
"significant" svn_rev change from a State_leave transition).

With this change, State_leave is no longer the transition that might result in Restart_server
and setting the new base revision. Instead, we switch to using Changed_merge_base instead.

Finally, this change adds a "pump watchman subscription until it is dry" during "make_report".
The reason for this is documented in the code, but to reiterate here: with this SCM aware mode,
Watchman is often polluting the pipe to our subscription with dozens of notifications after
and hg update. We used to read one message from this subscription per tick (one second), which meant
we wouldn't see the notifications in real time and new notifications that came shortly after
the batch would take 10+ seconds to be observed. By "pump it until it is dry", we non-blockingly
consume all the watchman notifications at once.

Reviewed By: dabek

Differential Revision: D5752843

fbshipit-source-id: bf517d862c7425e7af05f7c187c1ec3dad9a0f39