avoid reconnect-battles (regression)
Summary:
This fixes a bug where, whenever you open two Nuclide windows, then they get locked in a perpetual battle each one stealing the hh_server persistent connection from the other in turn.

(This also arises when your laptop shuts down but the nuclide-server lives on, and then you start up your laptop and launch a fresh instance of Nuclide, and it ends up with one defunct clientLsp instance on the server in addition to the new one. It surprises me that Nuclide allows this scenario, and I'll dig into that later.)

The bug was a regression introduced by a recent diff
https://github.com/facebook/hhvm/commit/e26ff2a5e0fc89688857313f284a54b7b15b5e44

Prior to that diff: if clientLsp loses its persistent connection to the server for reason of unexpected termination, but the server-monitor is still running (e.g. if you tweak .hhconfig, or if you do `hh restart`), then it will pop up "hh_server is stopped" for just a moment. On the next tick, it sees that the monitor lockfile is present, and so automatically reconnects.

But if clientLsp loses its persistent connection for reason that another Nuclide window has stolen the persistent connection, then the "tick" reconnect logic declines to be triggered by the monitor lockfile presence.

...

After that diff: as a shortcut, it checks for the monitor lockfile presence PRIOR to popping up "hh_server is stopped". If it sees the lockfile is present, then it reconnects immediately. This avoided the brief message flicker in the case of unexpected termination.

But the shortcut logic was being fired even in cases where another Nuclide window has stolen the persistent connection. It shouldn't have been. It should have used the same logic as the "tick" handler was already using.

This diff fixes the regression.

It also changes one related area, where it will only attempt to regain the persistent connection in response to an LSP request or notification. (Previously it would attempt to regain in response to an LSP request or notification or response. I don't think that a response merits an attempt to reconnect, since a response doesn't indicate a user-initiated action. For instance if clientLsp had put a dialog up, and then cancelled the dialog, the client would send an LSP response saying "the dialog has been closed". This shouldn't trigger a reconnect).

Reviewed By: arxanas

Differential Revision: D5804064

fbshipit-source-id: 7ca567631dad23a2cc1aa443da10be70954abbf3